sentiment analysis added. toggle disables browse btn
