--- gui/analysisReport.kv
@@ -1,79 +1,0 @@
-#:kivy 2.0.0
-#:import utils kivy.utils
-
-<AnalysisReport>:
-    # LAYOUT #
-    BoxLayout:
-        orientation: 'vertical'
-
-        # TITLE #
-        Label:
-            text: 'Feedback Response Analysis Report'
-            size_hint_y: None
-            height: 64
-            padding: (8, 8)
-
-        # CONTENT #
-        BoxLayout:
-            orientation: 'horizontal'
-
-            # GRID #
-            GridLayout:
-                cols: 2
-
-                # SUMMARY #
-                BoxLayout:
-                    orientation: 'vertical'
-                    Label:
-                        text: 'summary'
-                    Label:
-                        text: 'content'
-                        size_hint_y: None
-                        height: 32
-                # CHART #
-                BoxLayout:
-                    orientation: 'vertical'
-                    Label:
-                        text: 'chart'
-                    Label:
-                        text: 'content'
-                        size_hint_y: None
-                        height: 32
-                # TOPICS1 #
-                BoxLayout:
-                    orientation: 'vertical'
-                    Label:
-                        text: 'topics1'
-                    Label:
-                        text: 'content'
-                        size_hint_y: None
-                        height: 32
-                # TOPICS2 #
-                BoxLayout:
-                    orientation: 'vertical'
-                    Label:
-                        text: 'topics2'
-                    Label:
-                        text: 'content'
-                        size_hint_y: None
-                        height: 32
-
-            # SIDEBAR #
-            BoxLayout:
-                orientation: 'vertical'
-                size_hint_x: None
-                width: 200
-           
-                Button:
-                    text: 'topic-1'
-                Button:
-                    text: 'topic-2'
-                Button:
-                    text: 'topic-3'
-                Button:
-                    text: 'topic-4'
-                Button:
-                    text: 'topic-5'
-                Button:
-                    text: 'again'
-
