--- kivy/kivy-tutorial.py
@@ -1,41 +1,0 @@
-"""
-=================================================
-Author: Drew Rinker
-
-Kivy tutorial
-=================================================
-"""
-from typing import Text
-import kivy
-
-from kivy.uix.label import Label
-from kivy.uix.gridlayout import GridLayout
-from kivy.uix.textinput import TextInput
-from kivy.app import App
-
-kivy.require('2.0.0')
-
-
-class Login(GridLayout):
-
-    def __init__(self, **kwargs):
-        super(Login, self).__init__(**kwargs)
-        self.cols = 2
-
-        self.add_widget(Label(text='User Name'))
-        self.username = TextInput(multiline=False)
-        self.add_widget(self.username)
-
-        self.add_widget(Label(text='Password'))
-        self.password = TextInput(password=True, multiline=False)
-        self.add_widget(self.password)
-
-
-class MyApp(App):
-
-    def build(self):
-        return Login()
-
-
-if __name__ == '__main__':
-    MyApp().run()
