--- kivy/pong.kv
@@ -1,33 +1,0 @@
-#:kivy 2.0.0
-
-<PongGame>:
-
-    canvas:
-
-        Color:
-            rgb: 0,1,0
-
-        Rectangle:
-            pos: self.center_x - 5, 0
-            size: 10, self.height
-
-    Label:
-        font_size: 70
-        center_x: root.width / 4
-        top: root.top - 50
-        text: "0"
-
-    Label:
-        font_size: 70
-        center_x: root.width * 3 / 4
-        top: root.top - 50
-        text: "0"
-
-    Button:
-        text: "Hannah sucks"
-        center_x: root.width / 4
-        top: root.top - 200
-        font_size: 10
-
-
-
